# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ailisher Mobile App API Documentation",
    "description": "Unified interactive API documentation compiled automatically from the Postman app developer collections.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://ailisher.diintech.com",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/clients/{clientId}/mobile/banners": {
      "get": {
        "summary": "Get Grouped Active Banners",
        "description": "Fetches all active banners for the client, grouped by placement ('top', 'medium', 'bottom') and sorted by their priority order. Image URLs are dynamically signed using Cloudflare R2 presigned URLs.",
        "tags": [
          "2. Mobile Banners Collection | Banners Management"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter clientId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          }
        }
      }
    },
...
    "/api/clients/{clientId}/mobile/reels/user": {
      "get": {
        "summary": "Get Reels Feed (User)",
        "description": "Fetches list of all reels published by the client, marking each object with 'isLiked' and 'isViewed' flags evaluated specifically for the authenticated student user.",
        "tags": [
          "5. Current Affairs & Video Reels Collection | 2. Mobile Video Reels Feed (Student)"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter clientId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          }
        }
      }
    },
...
    "/api/reels": {
      "post": {
        "summary": "Publish a new Reel",
        "description": "Creates a new Reel. Admin must either provide the uploaded R2 'videoKey' or a valid 'youtubeLink'.",
        "tags": [
          "5. Current Affairs & Video Reels Collection | 3. Reels Feed Administration"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Union Budget 2026-27 Highlights"
                  },
                  "description": {
                    "type": "string",
                    "example": "A detailed analysis of sector budgets and tax alterations."
                  },
                  "videoKey": {
                    "type": "string",
                    "example": "my-business-client/reels/1782371852722.mp4"
                  },
                  "youtubeLink": {
                    "type": "string",
                    "example": "https://www.youtube.com/shorts/abcd1234efg"
                  }
                },
                "required": [
                  "title",
                  "description",
                  "videoKey",
                  "youtubeLink"
                ]
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Published Reels List (Admin)",
        "description": "Admin list of all reels created by the business client account. Automatically generates refreshed GET presigned URLs for R2 storage.",
        "tags": [
          "5. Current Affairs & Video Reels Collection | 3. Reels Feed Administration"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response"
          }
        }
      }
    },
...
    "/api/clients/{clientId}/mobile/books": {
      "get": {
        "summary": "Get Bookstore Books List",
        "description": "Queries all published books for the client with optional filters and keyword searches.",
        "tags": [
          "8. Bookstore & Cart Collection | 1. Mobile Bookstore Discovery"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter clientId"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "1"
            },
            "description": "Query parameter page"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "10"
            },
            "description": "Query parameter limit"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            },
            "description": "Query parameter search"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            },
            "description": "Query parameter category"
          },
          {
            "name": "subcategory",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            },
            "description": "Query parameter subcategory"
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            },
            "description": "Query parameter language"
          },
          {
            "name": "author",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            },
            "description": "Query parameter author"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            },
            "description": "Query parameter tag"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          }
        }
      }
    },
...
    "/api/clients/{clientId}/mobile/cart": {
      "get": {
        "summary": "Get Cart Items",
        "description": "Fetches current checkout items and total value inside the student's cart.",
        "tags": [
          "8. Bookstore & Cart Collection | 3. Shopping Cart Management"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter clientId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          }
        }
      }
    },
...
  }
}